Implement stubbed_attributes file
This is a redo of #28 to avoid merge conflicts between forks.

Add implementation of the stubbed_attributes.yml method of setting up a
stub session. This feature has already been documented in the
[README](https://github.umn.edu/asrweb/umn_shib_auth/tree/d12b14b2a976d6fc00824d1911e69ffe0abe75f6#configuration-file)

In regards to the controller_methods_spec, we now have multiple way to
stub the session, so this spec was failing because we were returning one
of the other sets of stubbed values - the sutbbed attributes yml.

This test doesn't care where they come from, so verify we get the values
from our collaborator, UmnShibAuth.