Use binstubs in development and only in dev
When we added binstubs to the project, we also added them to the
executables we were exporting with the gem.  Per [the specification](https://guides.rubygems.org/specification-reference/#executables),
the `executables` property is for executables associated directly with
your gem - for example, `rake` for the `rake` gem.  None of our binstubs
really match that definition, so it doesn't make sense to export them
with the gem.  This is also producing regular errors messages from
projects that use umn_shib_auth:

```
The `rake` executable in the `umn_shib_auth` gem is being loaded, but it's also present in other gems (rake).
If you meant to run the executable for another gem, make sure you use a project specific binstub (`bundle binstub <gem_name>`).
If you plan to use multiple conflicting executables, generate binstubs for them and disambiguate their names.
The `rake` executable in the `rake` gem is being loaded, but it's also present in other gems (umn_shib_auth).
If you meant to run the executable for another gem, make sure you use a project specific binstub (`bundle binstub <gem_name>`).
If you plan to use multiple conflicting executables, generate binstubs for them and disambiguate their names
```

This also changes the scripts to use the binstubs that we are
generating, rather than using `bundle exec`.