Add bundler-audit
Adds `bundler-audit` in order to scan for upgrades and security updates.
At the moment this is running in `script/setup` so that when you first
start working on the project you will be notified of any gem updates
that need to happen.  However, there is a time cost with this, since the
scan takes several minutes to run.  It may be better to add this to a
separate script that is run on an ad hoc basis.

Co-authored-by: Nikoli Hovorka <hovor009@umn.edu>
Co-authored-by: Shawn Isenhart <isen0011@umn.edu>